Unify server response object format
Unified the server response object to include status, message and data.
Modified client side to consume new format. Fixed a bug on logout when no
socket is opened - when logout is attempted without having at least one
game uploaded it fails with a server error and does not perform logout.